Remove SolidusBraintree::Spree namespace
This namespace very effectively hinders access to the root ::Spree
namespace and would make require every reference to the `Spree` constant
to be fully qualified.

Fixes the following build error:
```
exist  config/initializers
   identical  config/initializers/solidus_braintree.rb
        rake  railties:install:migrations FROM=solidus_braintree
         run  bin/rails db:migrate from "."
        gsub  config/routes.rb
/lib/generators/solidus_braintree/install/install_generator.rb:153:in `solidus_mount_point': uninitialized constant SolidusBraintree::Spree::Core (NameError)

        mount_point = Spree::Core::Engine.routes.find_script_name({})
                                 ^^^^^^^^
```
